Add "minification" utility functions
When building the loaders for both webpack and rollup, it was desireable
to limit the output of the AST for the svg components to only those
fields that were necessary to render the final output.  Instead of
maintaining these minification functions across implementations and
needing to track them with the requirements of this module, I'm adding
them as optional imports to this.